+++ chia/_tests/core/test_merkle_set.py
@@ -12,9 +12,9 @@ from chia_rs import Coin, MerkleSet, compute_merkle_set_root, confirm_included_a
 
 from chia.simulator.block_tools import BlockTools
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.util.batches import to_batches
 from chia.util.hash import std_hash
 from chia.util.ints import uint64
-from chia.util.misc import to_batches
 from chia.wallet.util.wallet_sync_utils import validate_additions, validate_removals
