+++ chia/_tests/wallet/test_wallet_blockchain.py
@@ -12,9 +12,9 @@ from chia.types.blockchain_format.vdf import VDFProof
 from chia.types.full_block import FullBlock
 from chia.types.header_block import HeaderBlock
 from chia.types.peer_info import PeerInfo
+from chia.util.batches import to_batches
 from chia.util.generator_tools import get_block_header
 from chia.util.ints import uint8, uint32
-from chia.util.misc import to_batches
 from chia.wallet.key_val_store import KeyValStore
 from chia.wallet.wallet_blockchain import WalletBlockchain
