+++ chia/full_node/hint_store.py
@@ -7,8 +7,8 @@ from typing import List, Set, Tuple
 import typing_extensions
 
 from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.util.batches import to_batches
 from chia.util.db_wrapper import SQLITE_MAX_VARIABLE_NUMBER, DBWrapper2
-from chia.util.misc import to_batches
 
 log = logging.getLogger(__name__)
